Move Scope to parser internal
